refactor: client 클래스명, 디렉토리, response 변경
