+++ src/main/java/com/coniverse/dangjang/domain/infrastructure/auth/client/NaverClient.java
@@ -1,4 +1,4 @@
-package com.coniverse.dangjang.domain.auth.service.oauthInfoRequest;
+package com.coniverse.dangjang.domain.infrastructure.auth.client;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpEntity;
@@ -10,39 +10,39 @@ import org.springframework.util.MultiValueMap;
 import org.springframework.web.client.RestTemplate;
 
 import com.coniverse.dangjang.domain.auth.dto.OauthProvider;
-import com.coniverse.dangjang.domain.user.infrastructure.KakaoInfoResponse;
-import com.coniverse.dangjang.domain.user.infrastructure.OAuthInfoResponse;
+import com.coniverse.dangjang.domain.infrastructure.auth.dto.NaverInfoResponse;
+import com.coniverse.dangjang.domain.infrastructure.auth.dto.OAuthInfoResponse;
 
 import lombok.RequiredArgsConstructor;
 
 /**
- * 카카오 사용자 정보 조회
+ * 네이버 사용자 정보 조회
  *
  * @author EVE
- * @since 1.0
+ * @since 1.0.0
  */
 @Component
 @RequiredArgsConstructor
-public class KakaoInfoRequestService implements OAuthInfoRequestService {
+public class NaverClient implements OAuthClient {
 
-	@Value("${oauth.kakao.url.api}")
+	@Value("${oauth.naver.url.api}")
 	private String apiUrl;
 
 	private final RestTemplate restTemplate;
 
 	@Override
 	public OauthProvider getOauthProvider() {
-		return OauthProvider.KAKAO;
+		return OauthProvider.NAVER;
 	}
 
 	/**
-	 * @param accessToken 카카오 accessToken을 받아온다.
-	 * @return OAuthInfoResponse 카카오 사용자 정보
-	 * @since 1.0
+	 * @param accessToken 네이버 accessToken
+	 * @return OAuthInfoResponse 네이버 사용자 정보
+	 * @since 1.0.0
 	 */
 	@Override
 	public OAuthInfoResponse requestOauthInfo(String accessToken) {
-		String url = apiUrl + "/v2/user/me";
+		String url = apiUrl + "/v1/nid/me";
 
 		HttpHeaders httpHeaders = new HttpHeaders();
 		httpHeaders.setContentType(MediaType.APPLICATION_FORM_URLENCODED);
@@ -52,6 +52,6 @@ public class KakaoInfoRequestService implements OAuthInfoRequestService {
 
 		HttpEntity<?> request = new HttpEntity<>(body, httpHeaders);
 
-		return restTemplate.postForObject(url, request, KakaoInfoResponse.class);
+		return restTemplate.postForObject(url, request, NaverInfoResponse.class);
 	}
 }
